+++ onsenui/esm/elements/ons-dialog.js
@@ -182,6 +182,14 @@ export default class DialogElement extends BaseDialogElement {
    *  [ja]背景のマスクの色を指定します。"rgba(0, 0, 0, 0.2)"がデフォルト値です。[/ja]
    */
 
+  /**
+   * @attribute visible
+   * @type {Boolean}
+   * @description
+   *   [en]Whether the dialog is visible or not.[/en]
+   *   [ja]要素が見える場合に`true`。[/ja]
+   */
+
   constructor() {
     super();
 
@@ -306,7 +314,6 @@ export default class DialogElement extends BaseDialogElement {
 
   /**
    * @property visible
-   * @readonly
    * @type {Boolean}
    * @description
    *   [en]Whether the dialog is visible or not.[/en]